backend/SSH_backend_final/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
fasttext-wheel>=0.9.2
spacy>=3.7.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.10.0
